+++ engine-dao/src/main/java/com/graduate/engine/model/TaskCheckPoint.java
@@ -4,13 +4,13 @@ import lombok.Data;
 
 @Data
 public class TaskCheckPoint {
-    private Integer taskCheckPointId;
+    private Long taskCheckPointId;
 
-    private Integer taskId;
+    private Long taskId;
 
     private String taskCheckPointName;
 
     private String taskCheckPointMemo;
 
-    private Integer taskCheckPointDate;
+    private Long taskCheckPointDate;
 }
